#32 split AppContextProvider into its own file to support React Fast Refresh
